refactor(sequences): move type pkgs to operation pkgs
Instead of having packages centered around types/metatypes, such as
slices, single, io, etc... have the centered around creating sequences,
or converting back from sequences.

Haven't touched the chans package yet, but refactored the others.
