+++ process/doc.go
@@ -1,0 +1,3 @@
+// Package process contains a host of functions for transforming sequences into
+// into other sequences.
+package process
